// actions/src/lib/github/api/repos/response/file-commit.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * File Commit
 */
export interface FileCommit {
  content: {
    name?: string;
    path?: string;
    sha?: string;
    size?: number;
    url?: string;
    html_url?: string;
    git_url?: string;
    download_url?: string;
    type?: string;
    _links?: {
      self?: string;
      git?: string;
      html?: string;
      [k: string]: unknown;
    };
    [k: string]: unknown;
  } | null;
  commit: {
    sha?: string;
    node_id?: string;
    url?: string;
    html_url?: string;
    author?: {
      date?: string;
      name?: string;
      email?: string;
      [k: string]: unknown;
    };
    committer?: {
      date?: string;
      name?: string;
      email?: string;
      [k: string]: unknown;
    };
    message?: string;
    tree?: {
      url?: string;
      sha?: string;
      [k: string]: unknown;
    };
    parents?: {
      url?: string;
      html_url?: string;
      sha?: string;
      [k: string]: unknown;
    }[];
    verification?: {
      verified?: boolean;
      reason?: string;
      signature?: string | null;
      payload?: string | null;
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  [k: string]: unknown;
}